requirements.txt
httpx
selectolax
pypdf
pydantic
pandas
pypdf
jinja2
python-dotenv
